document and remove message-id header
It is necessary for message-id to be unignored by mutt, or it won't be
made available to display_filter. As a consequence, also remove
message-id after it's used, so that it is not actually displayed.

I tried renaming it to i.e. X-Message-ID, but there's no way to ignore
the new header (afaict) so long as it has 'Message-ID' in its name.

It's unfortunate to just remove Message-ID. Hopefully it won't cause
confusion later.

Signed-off-by: Dan Rue <dan.rue@linaro.org>