Try to gather some kernelci build times
The results are problematic because the data is problematic; this script
attempts to correlate stable rc announcements with build times from
kernelci. The problem is that the tree often gets pushed multiple times.
Patch counts are used to help here, but even then, sometimes in the
worst case, stable-rc will be pushed 2 days before it's announced. Then
it's pushed with the same patch count and announced. and then it's still
possible to have another subsequent push again with the same number of
patches. In this worse case, this script would report the same stable
version was built in -2 days, N hours, and M hours, because the only
timestamps we have are the -rc email announcement and the kernelci build
reports.

Signed-off-by: Dan Rue <dan.rue@linaro.org>